move block tests into separate file
